// packages/easy-fm/src/types.ts
export interface databaseOptionsBase<
    CREDENTAILS =
        | loginOptionsOAuth
        | loginOptionsFileMaker
        | loginOptionsClaris
        | loginOptionsToken,
> {
    database: string;
    debug?: boolean;
    credentials: CREDENTAILS;
}

/** Database options including any external FileMaker sources. */
export interface databaseOptionsWithExternalSources<
    CREDENTIALS =
        | loginOptionsOAuth
        | loginOptionsFileMaker
        | loginOptionsClaris
        | loginOptionsToken,
> extends databaseOptionsBase<CREDENTIALS> {
    database: string;
    debug?: boolean;
    credentials: CREDENTIALS;
    externalSources: Array<databaseOptionsBase<loginOptionsFileMaker>>;
}

/** Use an existing FileMaker Data API token. */
export interface loginOptionsToken {
    method: 'token';
    token: string;
}

/** Use FileMaker Data API OAuth headers to create session. */
export interface loginOptionsOAuth {
    method: 'oauth';
    oauth: {
        requestId: string;
        requestIdentifier: string;
    };
}

/** Use standard FileMaker username/password login. */
export interface loginOptionsFileMaker {
    method: 'filemaker';
    username: string;
    password: string;
    /**
     * sessionPoolSize - The maximum number of sessions to keep open in the pool. Default is 4.
     */
    sessionPoolSize?: number;
}

/** Use Claris authentication for session creation. */
export interface loginOptionsClaris {
    method: 'claris';
    claris: {
        fmid: string;
    };
}
// ...
export const FMHostMetadata = z.object({
    productInfo: z.object({
        buildDate: z.coerce.date().optional(),
        name: z.string(),
        version: z.string().optional(),
        dateFormat: z.string(),
        timeFormat: z.string(),
        timeStampFormat: z.string(),
    }),
});